jartool.c (consume): Changed amt and tc to size_t.

+++ jartool.c
@@ -108,7 +108,7 @@ static void version(void);
 static void add_entry(struct zipentry *);
 static void init_headers(void);
 
-static int consume(pb_file *, int);
+static int consume(pb_file *, size_t);
 static int list_jar(int, const char**, int);
 static int extract_jar(int, const char**, int);
 static int add_file_to_jar(int, int, const char*, struct stat*, int);
@@ -2118,8 +2118,8 @@ int list_jar(int fd, const char **files, int file_num){
   return 0;
 }
 
-int consume(pb_file *pbf, int amt){
-  int tc = 0; /* total amount consumed */
+int consume(pb_file *pbf, size_t amt){
+  size_t tc = 0; /* total amount consumed */
   ub1 buff[RDSZ];
   size_t rdamt;
 
@@ -2128,7 +2128,7 @@ int consume(pb_file *pbf, int amt){
 #endif
 
   if (seekable){
-    if (amt <= (int)pbf->buff_amt)
+    if (amt <= pbf->buff_amt)
       pb_read(pbf, buff, amt);
     else {
       lseek(pbf->fd, amt - pbf->buff_amt, SEEK_CUR);
